LayoutTests/imported/w3c/web-platform-tests/gamepad/gamepad-secure-context-expected.txt

CONSOLE MESSAGE: Navigator.getGamepads() will be removed from insecure contexts in a future release

FAIL Gamepad API is not exposed in non-secure contexts assert_equals: Gamepad type is undefined expected "undefined" but got "function"
